move mrd back to where it was
